% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fastqc_modules.R
\name{get_fastqc_modules}
\alias{get_fastqc_modules}
\title{List available fastqc analysis modules and their reported status}
\usage{
get_fastqc_modules(file, report_linenumbers = F)
}
\arguments{
\item{file}{Path to fastqc_data.txt file}

\item{report_linenumbers}{FALSE by default}
}
\value{
A data_frame of modules available in a fastqc_data.txt report and their report status
}
\description{
\code{get_fastqc_modules} lists avaialble analysis modules from a fastqc report
}
\examples{
# get_fastqc_modules("fastqc_data.txt")
}

