% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastqc.R
\name{read_fastqc}
\alias{read_fastqc}
\title{Read in a fastqc_data.txt file.}
\usage{
read_fastqc(file, module = "Basic Statistics")
}
\arguments{
\item{file}{Path to fastqc_data.txt file}

\item{module}{Module to extract from fastqc_data.txt file}
}
\value{
A data_frame with fastqc report values for the queried module
}
\description{
\code{readFastQC} parses a fastqc_data.txt file reported by fastqc,
a quality control tool for high throughput sequence data.
}
\examples{
# read_fastqc("fastqc_data.txt", "Sequence Length Distribution")
}

